requirements.txt
discord
pytesseract
pillow
